% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_rankchange_analysis.R
\name{RCA}
\alias{RCA}
\title{extract differetially ranked genes for each subtype}
\usage{
RCA(data, sampAnnot)
}
\arguments{
\item{data}{data which you want to extract genes from;}

\item{sampAnnot}{sampAnnot which is the annotation file for data}
}
\value{
a global object all_rank_t_genes
}
\description{
this function is used to extratc differetially ranked genes from dataset for each subtype.
}
\examples{
RCA(GSE85217, sampAnnote_GSE85217)
}
